+++ Sources/StreamChatSwiftUI/ChatMessageList/MessageItemView.swift
@@ -144,8 +144,6 @@ public struct MessageItemView<Factory: ViewFactory>: View {
         )
         .accessibilityElement(children: .contain)
         .accessibilityIdentifier("MessageItemView")
-        // TODO: Refactor so LinkDetectionTextView does not depend directly on the view model through @Environment.
-        .environment(\.messageViewModel, messageViewModel)
         .onChange(of: message) { message in messageViewModel.message = message }
         .onChange(of: channel) { channel in messageViewModel.channel = channel }
     }
@@ -214,6 +212,8 @@ struct SwipeToReplyModifier: ViewModifier {
     let isSwipeToQuoteReplyPossible: Bool
     @Binding var quotedMessage: ChatMessage?
 
+    @Environment(\.layoutDirection) private var layoutDirection
+
     @Injected(\.images) private var images
     @Injected(\.utils) private var utils
     @Injected(\.colors) private var colors
@@ -240,6 +240,10 @@ struct SwipeToReplyModifier: ViewModifier {
     }
 
     private let feedbackGenerator = UIImpactFeedbackGenerator(style: .medium)
+
+    private var isRTL: Bool {
+        layoutDirection == .rightToLeft
+    }
 
     func body(content: Content) -> some View {
         content
@@ -292,7 +296,7 @@ struct SwipeToReplyModifier: ViewModifier {
                         .padding(.all, tokens.spacingXs)
                         .background(colors.buttonSecondaryBackground.toColor)
                         .clipShape(Circle())
-                        .offset(x: min(offsetX / 2, 50))
+                        .offset(x: min(offsetX / 2, 50) + (message.isRightAligned ? 30 : 0))
                     Spacer()
                 } : nil
             )
@@ -307,7 +311,7 @@ struct SwipeToReplyModifier: ViewModifier {
     }
 
     private func dragChanged(to value: CGFloat) {
-        let horizontalTranslation = value
+        let horizontalTranslation = isRTL ? -value : value
 
         if horizontalTranslation < 0 {
             return