+++ Sources/StreamChatSwiftUI/Utils.swift
@@ -75,6 +75,7 @@ import StreamChatCommonUI
     
     var _audioPlayer: AudioPlaying?
     var _audioRecorder: AudioRecording?
+    var linkDetector = TextLinkDetector()
     var pollsDateFormatter: PollTimestampFormatter = DefaultPollTimestampFormatter()
 
     public init(