+++ StreamChatSwiftUITests/Tests/Utils/ViewFactory_Tests.swift
@@ -121,7 +121,8 @@ import XCTest
                 message: message,
                 isFirst: true,
                 availableWidth: 300,
-                scrolledId: .constant(nil)
+                scrolledId: .constant(nil),
+                translationLanguage: nil
             )
         )
 
@@ -139,7 +140,8 @@ import XCTest
                 message: message,
                 isFirst: true,
                 availableWidth: 300,
-                scrolledId: .constant(nil)
+                scrolledId: .constant(nil),
+                translationLanguage: nil
             )
         )
 
@@ -983,21 +985,21 @@ import XCTest
         XCTAssert(view is MediaViewer<DefaultViewFactory>)
     }
     
-    func test_viewFactory_makeMediaViewerHeader() {
+    func test_viewFactory_makeMediaViewerToolbarModifier() {
         // Given
         let viewFactory = DefaultViewFactory.shared
-        
+
         // When
-        let view = viewFactory.makeMediaViewerHeader(
-            options: MediaViewerHeaderOptions(
+        let modifier = viewFactory.makeMediaViewerToolbarModifier(
+            options: MediaViewerToolbarModifierOptions(
                 title: .unique,
                 subtitle: .unique,
-                shown: .constant(true)
+                isShown: .constant(true)
             )
         )
-            
+
         // Then
-        XCTAssert(view is MediaViewerHeader)
+        XCTAssert(modifier is MediaViewerToolbarModifier)
     }
     
     func test_viewFactory_makeVideoPlayerHeaderView() {
@@ -1017,20 +1019,20 @@ import XCTest
         XCTAssert(view is MediaViewerHeader)
     }
     
-    func test_viewFactory_makeAddUsersView() {
+    func test_viewFactory_makeMemberAddView() {
         // Given
         let viewFactory = DefaultViewFactory.shared
         
         // When
-        let view = viewFactory.makeAddUsersView(
-            options: AddUsersViewOptions(
+        let view = viewFactory.makeMemberAddView(
+            options: MemberAddViewOptions(
                 options: .init(loadedUserIds: []),
                 onConfirm: { _ in }
             )
         )
         
         // Then
-        XCTAssert(view is AddUsersView<DefaultViewFactory>)
+        XCTAssert(view is MemberAddView<DefaultViewFactory>)
     }
 
     func test_viewFactory_makeStreamTextView() {
@@ -1038,7 +1040,7 @@ import XCTest
         let viewFactory = DefaultViewFactory.shared
 
         // When
-        let view = viewFactory.makeStreamTextView(options: .init(message: message))
+        let view = viewFactory.makeStreamTextView(options: .init(message: message, translationLanguage: nil))
 
         // Then
         XCTAssert(view is StreamTextView)
@@ -1049,7 +1051,7 @@ import XCTest
         let viewFactory = DefaultViewFactory.shared
 
         // When
-        let view = viewFactory.makeAttachmentTextView(options: .init(message: message, availableWidth: 300))
+        let view = viewFactory.makeAttachmentTextView(options: .init(message: message, availableWidth: 300, translationLanguage: nil))
 
         // Then
         XCTAssert(view is AttachmentTextView<DefaultViewFactory>)